refactor: add installation of NPM packages

+++ linux.sh
@@ -1,10 +1,5 @@
 #!/usr/bin/env bash
 
-RESOLUTION_W=1710
-RESOLUTION_H=1112
-RESOLUTION_FREQ=60
-DESKTOP_OUTPUT_NAME='Virtual-1'
-
 set -e
 
 echo 'Install packages.'
@@ -18,12 +13,41 @@ sudo apt install -y zoxide
 sudo apt install -y hunspell
 sudo apt install -y htop
 
-# Vim-plug(plugin manager)
+echo 'Set kitty as default terminal.'
+echo 1 | sudo update-alternatives --config x-terminal-emulator
+
+# Vim-plug(vim/nvim plugin manager)
 sh -c 'curl -fLo "${XDG_DATA_HOME:-$HOME/.local/share}"/nvim/site/autoload/plug.vim --create-dirs \
 	       https://raw.githubusercontent.com/junegunn/vim-plug/master/plug.vim'
 
-echo 'Set kitty as default terminal.'
-echo 1 | sudo update-alternatives --config x-terminal-emulator
+# Install Node Version Manager, Node(NPM included) and LSP's
+curl -o- https://raw.githubusercontent.com/nvm-sh/nvm/v0.40.3/install.sh | bash
+echo 'Assert that $PATH contains .nvm/../bin'
+echo $PATH
+read -r -p 'Continue? [Y/n]'; answer
+if [ "$answer" != 'Y' ]; then
+    exit 1
+fi
+#export NVM_DIR="$([ -z "${XDG_CONFIG_HOME-}" ] && printf %s "${HOME}/.nvm" || printf %s "${XDG_CONFIG_HOME}/nvm")"
+#[ -s "$NVM_DIR/nvm.sh" ] && \. "$NVM_DIR/nvm.sh"
+
+nvm install latest
+npm i -g bash-language-server
+npm i -g pyright
+npm i -g typescript-language-server typescript
+npm i -g vscode-langservers-extracted
+
+
+# Install Go
+curl https://go.dev/dl/go1.24.5.linux-arm64.tar.gz
+rm -rf /usr/local/go && tar -C /usr/local -xzf go1.24.5.linux-arm64.tar.gz
+echo 'Assert that $PATH contains /usr/local/go/bin'
+echo $PATH
+read -r -p 'Continue? [Y/n]'; answer
+if [ "$answer" != 'Y' ]; then
+    exit 1
+fi
+go install golang.org/x/tools/gopls@latest
 
 echo 'Generate SSH keys.'
 echo -n 'Insert comment for SSH keys: '; read ssh_comment
@@ -39,7 +63,12 @@ until [ "$answer" == "y" ]; do
     echo -n 'Are you done?[y/N] '; read answer
 done
 
-git clone git@github.com:marinacompsci/dotfiles.git $HOME/developer
+read -r -p 'Clone dotfiles now? [Y/n]' answer
+if [ "$answer" != 'Y' ]; then
+    exit 0
+fi
+
+git clone git@github.com:marinacompsci/dotfiles.git $HOME/developer/dotfiles
 
 rm -rf $HOME/.bash* $HOME/.tmux* $HOME/.vimrc
 rm -rf $HOME/.hunspell_en_US $HOME/.vim $HOME/.config/nvim/init.lua
@@ -52,22 +81,3 @@ SYMLINK_SCRIPT="$HOME/developer/dotfiles/scripts/bash/setup.sh"
 
 echo 'Install VM tools to help adjust the resolution.'
 sudo apt install open-vm-tools open-vm-tools-desktop
-
-echo "Calculate screen's resolution."
-CVT_OUTPUT=$(cvt $RESOLUTION_W $RESOLUTION_H $RESOLUTION_FREQ)
-NEW_MODELINE=$(echo $CVT_OUTPUT | sed -E 's/.*Modeline\s//')
-NEW_MODE=$(echo $NEW_MODELINE| sed -E "s/\"(.*)\".*/\1/")
-
-DISPLAY_SCRIPT="$HOME/linux/display.sh"
-cat > $DISPLAY_SCRIPT << EOF
-#!/usr/bin/env bash
-
-xrandr --newmode $NEW_MODELINE 2>&1 | logger -t i3-xrandr
-sleep 2 && xrandr --addmode $DESKTOP_OUTPUT_NAME $NEW_MODE 2>&1 | logger -t i3-xrandr
-sleep 2 && xrandr --output $DESKTOP_OUTPUT_NAME --mode $NEW_MODE 2>&1 | logger -t i3-xrandr
-EOF
-cat >> $HOME/.config/i3/config << EOF
-# Setting screen's dimensions and resolution.
-# Set this below the line where the font is set.
-exec --no-startup-id $DISPLAY_SCRIPT
-EOF
